tienda_virtual/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
mysqlclient==2.2.4
pip==24.0
setuptools==69.5.1
sqlparse==0.5.0
tzdata==2024.1
wheel==0.43.0